fix id generation bug

+++ src/main/java/com/myserver/portfolio/PortfolioController.java
@@ -20,7 +20,7 @@ public class PortfolioController {
     }
 
     @PostMapping(path = "create")
-    public Long addPortfolio(@RequestParam("name") String name){
+    public int addPortfolio(@RequestParam("name") String name){
         return portfolioService.create(name);
     }
 }
